refactor(common): replace java.time.LocalDateTime with kotlinx.datetime

+++ common/src/main/kotlin/dev/yidafu/blog/common/modal/BaseModel.kt
@@ -1,13 +1,13 @@
 package dev.yidafu.blog.common.modal
 
 import jakarta.persistence.*
-import java.time.LocalDateTime
+import kotlinx.datetime.LocalDateTime
 
 @MappedSuperclass
 open class BaseModel {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  var id: Long? = null
+  var id: Int? = null
 
   @Column(name = "created_at", columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
   var createdAt: LocalDateTime? = null
